+++ playground/backend/app/controllers/auth/session_controller.ts
@@ -2,7 +2,7 @@ import { handler } from '@folie/castle/helpers'
 
 export default class Controller {
   handle = handler(async ({ ctx }) => {
-    const user = await ctx.auth.session.getUser()
+    const user = await ctx.auth.session.getOwner()
 
     return { session: user.$serialize() }
   })