+++ playground/backend/app/controllers/note/create_controller.ts
@@ -4,7 +4,7 @@ import { handler } from '@folie/castle/helpers'
 
 export default class Controller {
   handle = handler(async ({ ctx }) => {
-    const user = await ctx.auth.session.getUser()
+    const user = await ctx.auth.session.getOwner()
 
     const metrics = await user.$metric().get()